Refactor nameable class
