+++ classes/decorator.rb
@@ -1,0 +1,14 @@
+require_relative './nameable'
+
+class Decorator < Nameable
+    attr_accessor :nameable
+  
+    def initialize(nameable)
+      @nameable = nameable
+      super()
+    end
+  
+    def correct_name
+      @nameable.correct_name
+    end
+  end
