+++ classes/nameable.rb
@@ -4,18 +4,7 @@ class Nameable
   end
 end
 
-class Decorator < Nameable
-  attr_accessor :nameable
 
-  def initialize(nameable)
-    @nameable = nameable
-    super()
-  end
-
-  def correct_name
-    @nameable.correct_name
-  end
-end
 
 class CapitalizeDecoration < Decorator
   def correct_name
